Multi-target for .NET 6 and MonoAndroid90
Context: https://github.com/xamarin/Xamarin.Legacy.Sdk

This uses Xamarin.Legacy.Sdk to multi-target.

I temporarily put a build of Xamarin.Legacy.Sdk in this commit. When
CI is green for this change, I'll put an official build on NuGet.org.

For this to work:

* You have to use `dotnet build` to build any `Xamarin.Legacy.Sdk`
  projects. Windows can use `MSBuild.exe` from Visual Studio 16.9
  Preview 4 as well.

The CI build will also have to install:

* .NET 6.0.100-alpha.1.21064.27
* Microsoft.NET.Android.Workload 11.0.200.43
* Xamarin.Android 11.2.99.43

Other notes:

* Class libraries had application properties set like
  `$(EnableProguard)`, `$(AndroidEnableMultiDex)`,
  `$(AndroidDexTool)`, `$(AndroidLinkTool)`, etc.